feat: 장바구니 선택 삭제 수정 및 supabase 서버 수정
