+++ src/app/(afterLogin)/home/lib/getMainDealList.ts
@@ -1,4 +1,4 @@
-import { createClient } from '@/app/utils/supabase/client';
+import { createClient } from '@/utils/supabase/client';
 
 export type GetMainDealListParams = {
   p_category_id?: number;
